Change BitField to use 64-bit int.

+++ scene/resources/font.cpp
@@ -264,7 +264,7 @@ Size2 Font::get_string_size(const String &p_text, HorizontalAlignment p_alignmen
 	hash = hash_djb2_one_64(p_font_size, hash);
 	if (p_alignment == HORIZONTAL_ALIGNMENT_FILL) {
 		hash = hash_djb2_one_64(hash_murmur3_one_float(p_width), hash);
-		hash = hash_djb2_one_64(p_jst_flags.operator uint32_t(), hash);
+		hash = hash_djb2_one_64(p_jst_flags.operator int64_t(), hash);
 	}
 	hash = hash_djb2_one_64(p_direction, hash);
 	hash = hash_djb2_one_64(p_orientation, hash);
@@ -293,8 +293,8 @@ Size2 Font::get_multiline_string_size(const String &p_text, HorizontalAlignment 
 	uint64_t hash = p_text.hash64();
 	hash = hash_djb2_one_64(p_font_size, hash);
 	hash = hash_djb2_one_64(hash_murmur3_one_float(p_width), hash);
-	hash = hash_djb2_one_64(p_brk_flags.operator uint32_t(), hash);
-	hash = hash_djb2_one_64(p_jst_flags.operator uint32_t(), hash);
+	hash = hash_djb2_one_64(p_brk_flags.operator int64_t(), hash);
+	hash = hash_djb2_one_64(p_jst_flags.operator int64_t(), hash);
 	hash = hash_djb2_one_64(p_direction, hash);
 	hash = hash_djb2_one_64(p_orientation, hash);
 
@@ -323,7 +323,7 @@ void Font::draw_string(RID p_canvas_item, const Point2 &p_pos, const String &p_t
 	hash = hash_djb2_one_64(p_font_size, hash);
 	if (p_alignment == HORIZONTAL_ALIGNMENT_FILL) {
 		hash = hash_djb2_one_64(hash_murmur3_one_float(p_width), hash);
-		hash = hash_djb2_one_64(p_jst_flags.operator uint32_t(), hash);
+		hash = hash_djb2_one_64(p_jst_flags.operator int64_t(), hash);
 	}
 	hash = hash_djb2_one_64(p_direction, hash);
 	hash = hash_djb2_one_64(p_orientation, hash);
@@ -359,8 +359,8 @@ void Font::draw_multiline_string(RID p_canvas_item, const Point2 &p_pos, const S
 	uint64_t hash = p_text.hash64();
 	hash = hash_djb2_one_64(p_font_size, hash);
 	hash = hash_djb2_one_64(hash_murmur3_one_float(p_width), hash);
-	hash = hash_djb2_one_64(p_brk_flags.operator uint32_t(), hash);
-	hash = hash_djb2_one_64(p_jst_flags.operator uint32_t(), hash);
+	hash = hash_djb2_one_64(p_brk_flags.operator int64_t(), hash);
+	hash = hash_djb2_one_64(p_jst_flags.operator int64_t(), hash);
 	hash = hash_djb2_one_64(p_direction, hash);
 	hash = hash_djb2_one_64(p_orientation, hash);
 
@@ -396,7 +396,7 @@ void Font::draw_string_outline(RID p_canvas_item, const Point2 &p_pos, const Str
 	hash = hash_djb2_one_64(p_font_size, hash);
 	if (p_alignment == HORIZONTAL_ALIGNMENT_FILL) {
 		hash = hash_djb2_one_64(hash_murmur3_one_float(p_width), hash);
-		hash = hash_djb2_one_64(p_jst_flags.operator uint32_t(), hash);
+		hash = hash_djb2_one_64(p_jst_flags.operator int64_t(), hash);
 	}
 	hash = hash_djb2_one_64(p_direction, hash);
 	hash = hash_djb2_one_64(p_orientation, hash);
@@ -432,8 +432,8 @@ void Font::draw_multiline_string_outline(RID p_canvas_item, const Point2 &p_pos,
 	uint64_t hash = p_text.hash64();
 	hash = hash_djb2_one_64(p_font_size, hash);
 	hash = hash_djb2_one_64(hash_murmur3_one_float(p_width), hash);
-	hash = hash_djb2_one_64(p_brk_flags.operator uint32_t(), hash);
-	hash = hash_djb2_one_64(p_jst_flags.operator uint32_t(), hash);
+	hash = hash_djb2_one_64(p_brk_flags.operator int64_t(), hash);
+	hash = hash_djb2_one_64(p_jst_flags.operator int64_t(), hash);
 	hash = hash_djb2_one_64(p_direction, hash);
 	hash = hash_djb2_one_64(p_orientation, hash);
